@@ -34,66 +35,67 @@
     "svelte": "^3.53.1 || ^4.0.5"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.5",
-    "@babel/core": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-typescript": "^8.2.1",
-    "@storybook/addon-actions": "^6.4.14",
-    "@storybook/addon-knobs": "^6.4.0",
-    "@storybook/addon-links": "^6.4.14",
+    "@babel/cli": "^7.23.0",
+    "@babel/core": "^7.23.2",
+    "@babel/preset-env": "^7.23.2",
+    "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-alias": "^5.0.1",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^25.0.5",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.5",
+    "@storybook/addon-actions": "^7.4.6",
+    "@storybook/addon-links": "^7.4.6",
     "@storybook/addon-options": "^5.3.21",
-    "@storybook/addon-storysource": "^6.4.14",
-    "@storybook/addons": "^6.4.14",
-    "@storybook/svelte": "^6.4.14",
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/svelte": "^3.0.3",
-    "@tsconfig/svelte": "^2.0.1",
-    "autoprefixer": "^9.8.6",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "clean-webpack-plugin": "^3.0.0",
-    "conventional-changelog-cli": "^2.1.1",
-    "copy-webpack-plugin": "^6.4.1",
+    "@storybook/addon-storysource": "^7.4.6",
+    "@storybook/addons": "^7.4.6",
+    "@storybook/svelte": "^7.4.6",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/svelte": "^4.0.4",
+    "@tsconfig/svelte": "^5.0.2",
+    "@types/jest": "^29.5.5",
+    "autoprefixer": "^10.4.16",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.1.3",
+    "clean-webpack-plugin": "^4.0.0",
+    "conventional-changelog-cli": "^4.1.0",
+    "copy-webpack-plugin": "^11.0.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.28.0",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-svelte3": "^3.2.0",
+    "eslint": "^8.51.0",
+    "eslint-plugin-jest": "^27.4.2",
+    "eslint-plugin-svelte3": "^4.0.0",
     "get-port-cli": "^3.0.0",
     "jest": "^29.7.0",
-    "jest-transform-svelte": "^2.1.1",
+    "jest-environment-jsdom": "^29.7.0",
     "json-loader": "^0.5.7",
-    "mini-css-extract-plugin": "^1.6.0",
+    "mini-css-extract-plugin": "^2.7.6",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.2.10",
-    "prettier": "^2.3.1",
-    "prettier-plugin-svelte": "^2.3.0",
-    "prismjs": "^1.25.0",
+    "postcss": "^8.4.31",
+    "prettier": "^3.0.3",
+    "prettier-plugin-svelte": "^3.0.3",
+    "prismjs": "^1.29.0",
     "raw-loader": "^4.0.2",
-    "rollup": "^2.51.1",
-    "rollup-plugin-alias": "^2.2.0",
+    "rollup": "^2.79.1",
     "rollup-plugin-analyzer": "^4.0.0",
-    "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-bundle-size": "^1.0.3",
-    "rollup-plugin-copy": "^3.4.0",
-    "rollup-plugin-livereload": "^2.0.0",
+    "rollup-plugin-copy": "^3.5.0",
+    "rollup-plugin-livereload": "^2.0.5",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-svelte": "^7.1.0",
+    "rollup-plugin-svelte": "^7.1.6",
     "rollup-plugin-terser": "^7.0.2",
-    "sirv-cli": "^1.0.12",
-    "standard-version": "^9.3.0",
-    "svelte": "^3.53.1 || ^4.0.5",
-    "svelte-check": "^2.9.2",
+    "sirv-cli": "^2.0.2",
+    "standard-version": "^9.5.0",
+    "svelte": "^3.59.2",
+    "svelte-check": "^3.5.2",
     "svelte-jester": "^3.0.0",
-    "svelte-loader": "^3.1.1",
-    "svelte-preprocess": "^4.7.3",
-    "typescript": "^4.3.2",
-    "wait-for-localhost-cli": "^2.0.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^4.11.1"
+    "svelte-loader": "^3.1.9",
+    "svelte-preprocess": "^5.0.4",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.2.2",
+    "wait-for-localhost-cli": "^3.0.0",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "dependencies": {
     "@popperjs/core": "^2.11.8"
